refactor: merged blockchain and p2p-specific code into a single package
