--- src/s2f/architecture/blockchain/package.ned
@@ -1,2 +1,0 @@
-package s2f.architecture.blockchain;
-@namespace(s2f::chain);
