--- src/s2f/architecture/blockchain/pow/package.ned
@@ -1,2 +1,0 @@
-package s2f.architecture.blockchain.pow;
-@namespace(s2f::chain::pow);
