+++ src/s2f/architecture/p2p/pow/chain/Block.h
@@ -12,7 +12,7 @@
 
 using namespace s2f::os::crypto;
 
-namespace s2f::chain::pow
+namespace s2f::p2p::pow
 {
 
     constexpr sha256digest GENESIS_HASH = {}; //<! Empty first block hash
