+++ src/s2f/architecture/p2p/pow/chain/InOut.h
@@ -6,7 +6,8 @@
 #include <omnetpp.h>
 
 using namespace s2f::os::crypto;
-namespace s2f::chain::pow
+
+namespace s2f::p2p::pow
 {
 
     /**
