+++ src/s2f/architecture/p2p/pow/chain/Transaction.cc
@@ -1,7 +1,7 @@
 #include "Transaction.h"
 #include "s2f/os/crypto/crypto.h"
 
-using namespace s2f::chain::pow;
+using namespace s2f::p2p::pow;
 using namespace s2f::os::crypto;
 
 Transaction Transaction::fromJSON(std::string data)
