+++ src/s2f/architecture/p2p/pow/nodes/tmp.txt
